Bugfixing

+++ app/Http/Controllers/OrganizationController.php
@@ -21,10 +21,12 @@ class OrganizationController extends Controller
      */
     public function index(): View
     {
-        Log::create([
-            'user_id' => Auth::user()->id,
-            'what' => 'organization.index page open/hover'
-        ]);
+        Log::create(
+            [
+                'user_id' => Auth::user()->id,
+                'what' => 'organization.index page open/hover'
+            ]
+        );
 
         return view('organization.index');
     }
@@ -34,10 +36,12 @@ class OrganizationController extends Controller
      */
     public function create(): View
     {
-        Log::create([
-            'user_id' => 1,
-            'what' => 'organization.create page open/hover'
-        ]);
+        Log::create(
+            [
+                'user_id' => 1,
+                'what' => 'organization.create page open/hover'
+            ]
+        );
         return view('organization.create');
     }
 
@@ -56,13 +60,12 @@ class OrganizationController extends Controller
                     'email' => $validated['email'],
                     'password' => $validated['password'],
                     'organization_id' => $validated['organization_id'],
-
                 ]
             );
             $user->assignRole('Servicer');
             $success = __('Successfully created an new employee.');
             DB::commit();
-            return redirect()->route('organizations.myorganizations')->with(compact('success'));
+            return redirect()->route('organizations.myorganization')->with(compact('success'));
         } catch (\Throwable $th) {
             DB::rollBack();
             return redirect()->back()->withInput()->with('error', $th->getMessage());
@@ -76,12 +79,14 @@ class OrganizationController extends Controller
     {
         DB::beginTransaction();
         try {
-            $request->validate([
-                'name' => 'required|string',
-                'city' => 'string',
-                'address' => 'string',
-                'zip' => 'string',
-            ]);
+            $request->validate(
+                [
+                    'name' => 'required|string',
+                    'city' => 'string',
+                    'address' => 'string',
+                    'zip' => 'string',
+                ]
+            );
             Organization::create(
                 [
                     'name' => $request->name,
@@ -90,10 +95,12 @@ class OrganizationController extends Controller
                     'zip' => $request->zip,
                 ]
             );
-            Log::create([
-                'user_id' => Auth::user()->id,
-                'what' => 'organization.create Organization created successfully |' . json_encode($request->all())
-            ]);
+            Log::create(
+                [
+                    'user_id' => Auth::user()->id,
+                    'what' => 'organization.create Organization created successfully |' . json_encode($request->all())
+                ]
+            );
             DB::commit();
             return redirect()->route('organizations.index')->with('success', __('Organization created successfully.'));
         } catch (\Throwable $th) {
@@ -115,33 +122,42 @@ class OrganizationController extends Controller
     }
     public function productMove(Request $request)
     {
-
         DB::beginTransaction();
         try {
             //validation
-            $request->validate([
-                'selected_user_id' => 'required',
-                'user_id' => 'required',
-                'product_id' => 'required',
-            ]);
+            $request->validate(
+                [
+                    'selected_user_id' => 'required',
+                    'user_id' => 'required',
+                    'product_id' => 'required',
+                ]
+            );
             $product = Product::with(['users'])->whereId($request->product_id)->first();
             $selectedUserId = $request->selected_user_id;
             $isAttached = $product->users->contains($selectedUserId);
             $product->users()->detach($request->user_id);
             $visible = Visible::whereUserId($request->user_id)->whereProductId($product->id)->first();
             if ($isAttached) {
-                Visible::whereUserId($request->selected_user_id)->whereProductId($product->id)->first()->update(['isVisible' => $visible->isVisible]);
+                Visible::whereUserId($request->selected_user_id)->whereProductId($product->id)->first()->update(
+                    [
+                        'isVisible' => $visible->isVisible
+                    ]
+                );
                 $visible->delete();
             } else {
                 $product->users()->attach($request->selected_user_id);
-                $visible->update(['user_id' => $request->selected_user_id]);
+                $visible->update(
+                    [
+                        'user_id' => $request->selected_user_id
+                    ]
+                );
             }
-
-
-            Log::create([
-                'user_id' => Auth::user()->id,
-                'what' => 'product successfully moved from user:' . $request->user_id . ' to user: ' . $request->selected_user_id
-            ]);
+            Log::create(
+                [
+                    'user_id' => Auth::user()->id,
+                    'what' => 'product successfully moved from user:' . $request->user_id . ' to user: ' . $request->selected_user_id
+                ]
+            );
 
             DB::commit();
             return redirect()->route('organizations.myorganization')->with('success', __('Product successfully moved.'));
@@ -186,7 +202,6 @@ class OrganizationController extends Controller
                     'zip' => $request->zip,
                 ]
             );
-
             DB::commit();
             return redirect()->route('organizations.index')->with('success', __('Organization updated successfully.'));
         } catch (\Throwable $th) {
@@ -202,9 +217,7 @@ class OrganizationController extends Controller
     {
         DB::beginTransaction();
         try {
-
             $organization->delete();
-
             DB::commit();
             return redirect()->route('organizations.index')->with('success', __('Organizations deleted successfully.'));
         } catch (\Throwable $th) {
@@ -216,13 +229,6 @@ class OrganizationController extends Controller
     {
 
         $user->products()->detach($product->id);
-        //dd($organization);
-        /*$organization = Organization::whereId($organization->id)->first();
-        $organization_id = $organization->id;
-        $products = Product::whereHas('users.organization', function ($query) use ($organization_id) {
-            $query->where('id', $organization_id);
-        })->get();
-        return view('organization.edit', compact('organization', 'products'));*/
         $user = Auth::user();
         $organization_id = $user->organization_id;
         $organization = Organization::with('users.products')->whereId($organization_id)->first();
@@ -240,13 +246,15 @@ class OrganizationController extends Controller
     {
         DB::beginTransaction();
         try {
-            $request->validate([
-                'name' => 'required|string',
-                'city' => 'string',
-                'address' => 'string',
-                'zip' => 'string',
-                'tax_number' => 'required|max:24',
-            ]);
+            $request->validate(
+                [
+                    'name' => 'required|string',
+                    'city' => 'string',
+                    'address' => 'string',
+                    'zip' => 'string',
+                    'tax_number' => 'required|max:24',
+                ]
+            );
             $organization->update(
                 [
                     'name' => $request->name,
@@ -267,16 +275,17 @@ class OrganizationController extends Controller
     {
         DB::beginTransaction();
         try {
-
-            $organization = Auth::user()->organization();
-            $organization->users()->detach($user);
-            if ($user->id != Auth::user()->id)
+            if ($user->id != Auth::user()->id && !$user->hasRole('Organizer')) {
+                $user->products()->detach();
+                $user->are_visible()->delete();
                 $user->delete();
-            DB::commit();
-            return redirect()->route('organizations.myorganization')->with('success', __('Successfully removed the user from your organization.'));
+                DB::commit();
+                return redirect()->route('organizations.myorganization')->with('success', __('Successfully removed the user from your organization.'));
+            }
+            return redirect()->route('organizations.myorganization')->with('error', __("The user could not be removed from the organisation. You cannot delete your account here."));
         } catch (\Throwable $th) {
             DB::rollback();
-            return redirect()->route('organizations.myorganization')->with('error', __("The user could not be removed from the organisation. You cannot delete your account here."));
+            return redirect()->route('organizations.myorganization')->with('error', __("The user could not be removed from the organisation. You cannot delete your account here." . $th->getMessage()));
         }
     }
 }
